Refactor experimental defer directive validation
