+++ apollo-ios-codegen/Sources/GraphQLCompiler/JavaScript/src/index.ts
@@ -10,11 +10,6 @@ import {
   buildASTSchema,
   printSchema,
   extendSchema,
-  Kind,
-  GraphQLDeferDirective,
-  GraphQLDirective,
-  DirectiveDefinitionNode,
-  ASTNode,
 } from "graphql";
 import { defaultValidationRules, ValidationOptions } from "./validationRules";
 import { compileToIR, CompilationResult } from "./compiler";
@@ -22,7 +17,7 @@ import { assertValidSchema, assertValidSDL } from "./utilities/graphql";
 import {
   addApolloCodegenSchemaExtensionToDocument,
 } from "./utilities/apolloCodegenSchemaExtension";
-import { definitionNode } from "./utilities";
+import { addExperimentalDeferDirectiveToDocument } from "./utilities/experimentalDeferDirective";
 
 // We need to export all the classes we want to map to native objects,
 // so we have access to the constructor functions for type checks.
@@ -57,7 +52,7 @@ export function loadSchemaFromSources(sources: Source[]): GraphQLSchema {
   }
 
   var document = addApolloCodegenSchemaExtensionToDocument(concatAST(documents))
-  document = validateOrAddExperimentalDeferDirectiveToDocument(document)
+  document = addExperimentalDeferDirectiveToDocument(document)
 
   if (!introspectionJSONResult) { assertValidSDL(document) }
 
@@ -111,52 +106,4 @@ export function compileDocument(
   validationOptions: ValidationOptions
 ): CompilationResult {
   return compileToIR(schema, document, legacySafelistingCompatibleOperations, validationOptions);
-}
-
-// While @defer is experimental the directive needs to be manually added to the list of available 
-// directives for the schema document. If the directive is already in the document it must be 
-// validated to ensure it matches the directive definition that the Apollo iOS defer feature was
-// built to support.
-//
-// Once defer is part of the GraphQL spec and the directive is no longer considered experimental
-// this check can be removed.
-function validateOrAddExperimentalDeferDirectiveToDocument(document: DocumentNode): DocumentNode {
-  const definition = document.definitions.find(isDeferDirectiveDefinitionNode)
-
-  if (definition === undefined) {
-    return concatAST([document, experimentalDeferDirectiveDocumentNode()])
-  }
-
-  if (!match(definition, GraphQLDeferDirective)) {
-    throw new Error(
-      `Mismatched ${GraphQLDeferDirective.toString()} directive found.`,
-    );
-  }
-
-  return document
-}
-
-// Returns whether the current node is a defer directive definition node.
-function isDeferDirectiveDefinitionNode(node: ASTNode): node is DirectiveDefinitionNode {
-  return (
-    node.kind === Kind.DIRECTIVE_DEFINITION && 
-    node.name.value === GraphQLDeferDirective.name
-  )
-}
-
-// Checks whether the supplied directive definition node matches against important properties
-// of the experimentally defined defer directive that Apollo iOS expects.
-function match(definition: DirectiveDefinitionNode, target: GraphQLDirective): Boolean {
-  return(
-    definition.repeatable === target.isRepeatable &&
-    definition.locations.map((node) => node.value).sort().toString() === target.locations.slice(0).sort().toString() &&
-    definition.arguments?.map((value) => value.name.value).sort().toString() === target.args.map((value) => value.name).sort().toString()
-  )
-}
-
-function experimentalDeferDirectiveDocumentNode(): DocumentNode {
-  return {
-    kind: Kind.DOCUMENT,
-    definitions: [definitionNode(GraphQLDeferDirective)]
-  }
 }
